moved relationship browser code into its own package
