+++ Core/src/org/sleuthkit/autopsy/communications/Bundle.properties
@@ -39,24 +39,9 @@ VisualizationPanel.organicLayoutButton.text=Organic
 VisualizationPanel.hierarchyLayoutButton.text=Hierarchical
 VisualizationPanel.clearVizButton.text_1=
 VisualizationPanel.snapshotButton.text_1=Snapshot Report
-ContactDetailsPane.nameLabel.text=jLabel1
 VisualizationPanel.clearVizButton.actionCommand=
 VisualizationPanel.backButton.toolTipText=Click to go back
 VisualizationPanel.forwardButton.toolTipText=Click to go forward
 VisualizationPanel.fastOrganicLayoutButton.toolTipText=Click to redraw the chart
 VisualizationPanel.clearVizButton.toolTipText=Click to clear the chart
 VisualizationPanel.forwardButton.text=
-SummaryViewer.countsPanel.border.title=Counts
-SummaryViewer.emailLabel.text=Emails:
-SummaryViewer.contactsLabel.text=Contacts:
-SummaryViewer.attachmentsLabel.text=Attachments:
-SummaryViewer.fileReferencesPanel.border.title=File References in Current Case
-SummaryViewer.caseReferencesPanel.border.title=Other Occurrences
-OutlineViewPanel.messageLabel.text=<Control Disabled>
-SummaryViewer.messagesDataLabel.text=messages
-SummaryViewer.callLogsDataLabel.text=callLogs
-SummaryViewer.contactsDataLabel.text=contacts
-SummaryViewer.emailDataLabel.text=emails
-SummaryViewer.attachmentsDataLabel.text=attachments
-SummaryViewer.messagesLabel.text=Messages:
-SummaryViewer.callLogsLabel.text=Call Logs:
