+++ Core/src/org/sleuthkit/autopsy/communications/Bundle.properties-MERGED
@@ -3,17 +3,6 @@ AccountNode.accountType=Type
 AccountNode.device=Device
 AccountNode.messageCount=Msgs
 applyText=Apply
-ContactNode_Email=Email Address
-ContactNode_Home_Number=Home Number
-ContactNode_Mobile_Number=Mobile Number
-ContactNode_Name=Name
-ContactNode_Office_Number=Office Number
-ContactNode_Phone=Phone Number
-ContactNode_URL=URL
-ContactsViewer_columnHeader_Email=Email
-ContactsViewer_columnHeader_Name=Name
-ContactsViewer_columnHeader_Phone=Phone
-ContactsViewer_tabTitle=Contacts
 CTL_OpenCVTAction=Communications
 CVTTopComponent.name=\ Communications Visualization
 CVTTopComponent.TabConstraints.tabTitle=Visualize
@@ -33,30 +22,12 @@ FiltersPanel.refreshButton.text=Refresh
 FiltersPanel.deviceRequiredLabel.text=Select at least one.
 FiltersPanel.accountTypeRequiredLabel.text=Select at least one.
 FiltersPanel.needsRefreshLabel.text=Displayed data is out of date. Press Refresh.
-MessageNode_Node_Property_Attms=Attachments
-MessageNode_Node_Property_Date=Date
-MessageNode_Node_Property_From=From
-MessageNode_Node_Property_Subject=Subject
-MessageNode_Node_Property_To=To
-MessageNode_Node_Property_Type=Type
-MessageViewer_columnHeader_Attms=Attachments
-MessageViewer_columnHeader_Date=Date
-MessageViewer_columnHeader_From=From
-MessageViewer_columnHeader_Subject=Subject
-MessageViewer_columnHeader_To=To
-MessageViewer_tabTitle=Messages
 OpenCVTAction.displayName=Communications
 PinAccountsAction.pluralText=Add Selected Accounts to Visualization
 PinAccountsAction.singularText=Add Selected Account to Visualization
 refreshText=Refresh Results
 ResetAndPinAccountsAction.pluralText=Visualize Only Selected Accounts
 ResetAndPinAccountsAction.singularText=Visualize Only Selected Account
-SummaryViewer_CaseRefNameColumn_Title=Case Name
-SummaryViewer_CentralRepository_Message=<Enable Central Resposity to see Case References>
-SummaryViewer_Creation_Date_Title=Creation Date
-SummaryViewer_FileRefNameColumn_Title=Path
-SummaryViewer_TabTitle=Summary
-ThumbnailViewer_Name=Media
 UnpinAccountsAction.pluralText=Remove Selected Accounts
 UnpinAccountsAction.singularText=Remove Selected Account
 VisalizationPanel.paintingError=Problem painting visualization.
@@ -111,27 +82,12 @@ VisualizationPanel.organicLayoutButton.text=Organic
 VisualizationPanel.hierarchyLayoutButton.text=Hierarchical
 VisualizationPanel.clearVizButton.text_1=
 VisualizationPanel.snapshotButton.text_1=Snapshot Report
-ContactDetailsPane.nameLabel.text=jLabel1
 VisualizationPanel.clearVizButton.actionCommand=
 VisualizationPanel.backButton.toolTipText=Click to go back
 VisualizationPanel.forwardButton.toolTipText=Click to go forward
 VisualizationPanel.fastOrganicLayoutButton.toolTipText=Click to redraw the chart
 VisualizationPanel.clearVizButton.toolTipText=Click to clear the chart
 VisualizationPanel.forwardButton.text=
-SummaryViewer.countsPanel.border.title=Counts
-SummaryViewer.emailLabel.text=Emails:
-SummaryViewer.contactsLabel.text=Contacts:
-SummaryViewer.attachmentsLabel.text=Attachments:
-SummaryViewer.fileReferencesPanel.border.title=File References in Current Case
-SummaryViewer.caseReferencesPanel.border.title=Other Occurrences
-OutlineViewPanel.messageLabel.text=<Control Disabled>
-SummaryViewer.messagesDataLabel.text=messages
-SummaryViewer.callLogsDataLabel.text=callLogs
-SummaryViewer.contactsDataLabel.text=contacts
-SummaryViewer.emailDataLabel.text=emails
-SummaryViewer.attachmentsDataLabel.text=attachments
-SummaryViewer.messagesLabel.text=Messages:
-SummaryViewer.callLogsLabel.text=Call Logs:
 VisualizationPanel_action_dialogs_title=Communications
 VisualizationPanel_action_name_text=Snapshot Report
 VisualizationPanel_module_name=Communications
