+++ Core/src/org/sleuthkit/autopsy/communications/relationships/AccountSourceContentChildNodeFactory.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package org.sleuthkit.autopsy.communications;
+package org.sleuthkit.autopsy.communications.relationships;
 
 import java.util.List;
 import java.util.Set;
