+++ Core/src/org/sleuthkit/autopsy/communications/relationships/Bundle.properties
@@ -1,0 +1,15 @@
+ContactDetailsPane.nameLabel.text=Placeholder
+SummaryViewer.countsPanel.border.title=Counts
+SummaryViewer.emailLabel.text=Emails:
+SummaryViewer.contactsLabel.text=Contacts:
+SummaryViewer.attachmentsLabel.text=Attachments:
+SummaryViewer.fileReferencesPanel.border.title=File References in Current Case
+SummaryViewer.caseReferencesPanel.border.title=Other Occurrences
+OutlineViewPanel.messageLabel.text=<Control Disabled>
+SummaryViewer.messagesDataLabel.text=messages
+SummaryViewer.callLogsDataLabel.text=callLogs
+SummaryViewer.contactsDataLabel.text=contacts
+SummaryViewer.emailDataLabel.text=emails
+SummaryViewer.attachmentsDataLabel.text=attachments
+SummaryViewer.messagesLabel.text=Messages:
+SummaryViewer.callLogsLabel.text=Call Logs:
