+++ Core/src/org/sleuthkit/autopsy/communications/relationships/Bundle.properties-MERGED
@@ -1,0 +1,44 @@
+ContactDetailsPane.nameLabel.text=Placeholder
+ContactNode_Email=Email Address
+ContactNode_Home_Number=Home Number
+ContactNode_Mobile_Number=Mobile Number
+ContactNode_Name=Name
+ContactNode_Office_Number=Office Number
+ContactNode_Phone=Phone Number
+ContactNode_URL=URL
+ContactsViewer_columnHeader_Email=Email
+ContactsViewer_columnHeader_Name=Name
+ContactsViewer_columnHeader_Phone=Phone
+ContactsViewer_tabTitle=Contacts
+MessageNode_Node_Property_Attms=Attachments
+MessageNode_Node_Property_Date=Date
+MessageNode_Node_Property_From=From
+MessageNode_Node_Property_Subject=Subject
+MessageNode_Node_Property_To=To
+MessageNode_Node_Property_Type=Type
+MessageViewer_columnHeader_Attms=Attachments
+MessageViewer_columnHeader_Date=Date
+MessageViewer_columnHeader_From=From
+MessageViewer_columnHeader_Subject=Subject
+MessageViewer_columnHeader_To=To
+MessageViewer_tabTitle=Messages
+SummaryViewer.countsPanel.border.title=Counts
+SummaryViewer.emailLabel.text=Emails:
+SummaryViewer.contactsLabel.text=Contacts:
+SummaryViewer.attachmentsLabel.text=Attachments:
+SummaryViewer.fileReferencesPanel.border.title=File References in Current Case
+SummaryViewer.caseReferencesPanel.border.title=Other Occurrences
+OutlineViewPanel.messageLabel.text=<Control Disabled>
+SummaryViewer.messagesDataLabel.text=messages
+SummaryViewer.callLogsDataLabel.text=callLogs
+SummaryViewer.contactsDataLabel.text=contacts
+SummaryViewer.emailDataLabel.text=emails
+SummaryViewer.attachmentsDataLabel.text=attachments
+SummaryViewer.messagesLabel.text=Messages:
+SummaryViewer.callLogsLabel.text=Call Logs:
+SummaryViewer_CaseRefNameColumn_Title=Case Name
+SummaryViewer_CentralRepository_Message=<Enable Central Resposity to see Case References>
+SummaryViewer_Creation_Date_Title=Creation Date
+SummaryViewer_FileRefNameColumn_Title=Path
+SummaryViewer_TabTitle=Summary
+ThumbnailViewer_Name=Media
