+++ Core/src/org/sleuthkit/autopsy/communications/relationships/MessagesChildNodeFactory.java
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.sleuthkit.autopsy.communications;
+package org.sleuthkit.autopsy.communications.relationships;
 
 import java.util.List;
 import java.util.Set;
