+++ Core/src/org/sleuthkit/autopsy/communications/relationships/SelectionInfo.java
@@ -16,13 +16,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.sleuthkit.autopsy.communications;
+package org.sleuthkit.autopsy.communications.relationships;
 
 import java.util.HashSet;
 import java.util.Set;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import org.openide.util.Exceptions;
 import org.sleuthkit.autopsy.casemodule.Case;
 import org.sleuthkit.autopsy.casemodule.NoCurrentCaseException;
 import org.sleuthkit.datamodel.Account;
@@ -54,7 +53,7 @@ public final class SelectionInfo {
      * @param accountDeviceInstances Selected accountDecivedInstances
      * @param communicationFilter    Currently selected communications filters
      */
-    SelectionInfo(Set<AccountDeviceInstance> accountDeviceInstances, CommunicationsFilter communicationFilter) {
+    public SelectionInfo(Set<AccountDeviceInstance> accountDeviceInstances, CommunicationsFilter communicationFilter) {
         this.accountDeviceInstances = accountDeviceInstances;
         this.communicationFilter = communicationFilter;
